// packages/typespec-ts/src/rlc-common/interfaces.ts
export enum SchemaContext {
  /** Schema is used as an input to an operation. */
  Input = "input",
  /** Schema is used as an output from an operation. */
  Output = "output",
  /** Schema is used as an exception from an operation. */
  Exception = "exception"
}

export interface Schema {
  name: string;
  type: string;
  typeName?: string;
  outputTypeName?: string;
  description?: string;
  required?: boolean;
  default?: any;
  readOnly?: boolean;
  usage?: SchemaContext[];
  alias?: string;
  outputAlias?: string;
  fromCore?: boolean;
  enum?: any[];
  isConstant?: boolean;
}

export interface ObjectSchema extends Schema {
  properties?: Record<string, Schema>;
  discriminatorValue?: string;
  discriminator?: Schema;
  isPolyParent?: boolean;
  isMultipartBody?: boolean;
  children?: {
    all?: ObjectSchema[];
    immediate?: ObjectSchema[];
  };
  parents?: {
    all?: ObjectSchema[];
    immediate?: ObjectSchema[];
  };
}

export interface DictionarySchema extends Schema {
  valueTypeName?: string;
  outputValueTypeName?: string;
  additionalProperties?: Schema;
}

export interface ArraySchema extends Schema {
  items?: Schema;
}